UTA-0069: read the lightmap clamps as the compact indices they are
The element is thirty bytes at its smallest, not fixed at thirty. UClamp
and VClamp are compact indices: 0x40 is the continue bit, so a clamp of
64 or more takes two bytes where 16 or 32 takes one. A lightmap's texel
dimensions are usually small, which is why a fixed thirty consumed 9560
exports before shifting every later element of any export holding a
large clamp.

Seen directly rather than swept -- in one MH-TheBoxWorld2 Model the
elements drift one byte at a time, and the element that starts the drift
reads `04 40 01` where its neighbours read `10 10`.

Over the install: consumed exactly 559932 -> 560276, refused 1420 ->
1076. Eight buckets are now empty -- lightmap entries, lightmap bytes,
bounds, leaf hulls, leaves count, lights, trailing fields, and completed
at the wrong offset. What is left is leaves 842 and package version
below 62 234, and both of those are refusals this reader makes on
purpose.

This does not reopen the ruling that UClamp and VClamp are not compact.
That was measured against SS 4.5's layout, where the clamps trail two
LEADING compact indices; it refuted the clamps being compact THERE, and
they are not. The offsets are the raw fields and the clamps are the
compact ones -- the section has the pair the wrong way round.

The fixture now writes a vClamp over 63, so a fixed-width read of the
clamps misaligns rather than agreeing by accident, and that mutation is
killed.

CLAUDE.md rule 14: no contract document was edited, so no gate ran.

+++ src/upkg/Geometry.h
@@ -142,15 +142,14 @@ struct ZoneProperties {
 /// This reader returns the table's indices and offsets; what the lightmap
 /// DATA means is a renderer question and out of scope (UTA-0069 SS 9).
 /// The members are in the order the file writes them, which is not the order
-/// SS 4.5 states. `dataOffset` and `iLightActors` are raw `i32`, and the two
-/// clamps are the lightmap's texel dimensions and are one byte each, so the
-/// element is a fixed thirty bytes rather than a variable twenty-eight-plus.
-/// UTA-0069's ROADMAP bullet carries the derivation and its corroboration.
+/// SS 4.5 states. That section has the compact fields the wrong way round:
+/// `dataOffset` and `iLightActors` are raw `i32`, and the two clamps are the
+/// compact ones. UTA-0069's ROADMAP bullet carries the derivation.
 struct LightMapIndex {
     std::int32_t dataOffset = 0; // an offset INTO lightBits, not a compact index
     Vector3 pan;
-    std::uint8_t uClamp = 0;
-    std::uint8_t vClamp = 0;
+    std::int32_t uClamp = 0; // compact, so an element is thirty bytes or more
+    std::int32_t vClamp = 0;
     float uScale = 0;
     float vScale = 0;
     std::int32_t iLightActors = 0; // -1 where the surface is lit by nothing
